apps/reviews/views.py
from rest_framework import generics
from rest_framework.permissions import IsAuthenticated, AllowAny
from django_filters.rest_framework import DjangoFilterBackend

from .models import Review
from .serializers import ReviewSerializer, ReviewCreateSerializer


class ReviewListCreateView(generics.ListCreateAPIView):
    """
    GET: Barcha fikrlar (?seller_id=X orqali filterlash)
    POST: Yangi fikr qoldirish (faqat autentifikatsiya qilingan)
    """
    filter_backends = [DjangoFilterBackend]
    filterset_fields = ['seller']

    def get_permissions(self):
        if self.request.method == 'POST':
            return [IsAuthenticated()]
        return [AllowAny()]

    def get_serializer_class(self):
        if self.request.method == 'POST':
            return ReviewCreateSerializer
        return ReviewSerializer

    def get_queryset(self):
        queryset = Review.objects.select_related(
            'reviewer', 'seller', 'order__product'
        )
        seller_id = self.request.query_params.get('seller_id')
        if seller_id:
            queryset = queryset.filter(seller_id=seller_id)
        return queryset
